Add support for new frontmatter boolean param, 'comment' to turn on/off Disqus embedding for the page.
By default, if Disqus is configured and you aren't on localhost, the
Disqus template will be inserted for pages with section "post" and
will not be inserted otherwise.  If comment is set to true, they'll be
inserted regardless of section, and if set to false will not be
inserted regardless of section.